Update window and recording UI
Updated window values default and minimums to prevent bad design when
the height is too small.

Updated the default width to be larger in development to compensate for
the devTools.

Added a Toast to the settings for the recordings location to show a
success message on successful change.

Update the Record page to use icons with the buttons and added padding
to the bottom.