View recordings functional
Update `tsconfig.json` by adding the `src/preload.d.ts` to the `files`
property so that it will be used in all typescript files.

Update `src/preload.d.ts`
- Added the channel type for `getRecordings`
- Added `Video` type for the recordings which consists of the filename
  and the buffer
- Added `getRecordings` method definition

Update `src/lib/permissionsHandler.ts` by adding the `fullscreen`
permission to allow fullscreen video playback on recordings.

Removed `src/main/menu.ts` because it was the original boilerplate code
for the windows frame but, we are using a custom Titlebar component to
handle it.

Update `src/main/window.ts` removing the references of building the menu
since we aren't using it but a custom one instead.

Update `src/main/ipc/videoStream.ts`
- Using `util.promisify()` on the `fs` methods to read and write files
- Added `timestamp` helper function to create easier to read time
  formats to saved recordings.
- Added an event listener on the `app` for the `ready` event to clear
  the currentStream id stored in case the app wasn't able to clear it
  when closed previously or crashed.
- Added the handler for the `getRecordings` event, which reads the
  recording directory stored in the config and returns an array of the
  videos or an error message if something failed.

Update `src/preload/preload.ts`
- Added the `getRecordings` implementation for the renderer on the
  `videoStream` API

Update `src/renderer/components/Layout.tsx`
- Added the `overflow: hidden` value to the wrapper to hide any excess
  content
- Added main class to the main element

Removed `src/renderer/components/Main.tsx because it was the original
prototype and is now obsolete and not needed.

Update `src/renderer/components/Sidebar.tsx adding the transition
property to the sidebar-button for the hover effect and others.

Update `src/renderer/components/Titlebar.tsx by changing the z-index
from `10` to `999` so that it will remain accessible despite the
overlay.

Update `src/renderer/pages/Recordings.tsx` to create a list of the
recordings and a custom modal to view the recording when clicked.