v1.0.0 Released working product
The Electron-Screen-Recorder is a portable app that simply records the
screen and can view previously saved recordings.

Updated `config/webpack.config.renderer.ts` CspHtmlWebpackPlugin so that
it disables nonce values because they are generated on the fly and
aren't added to the CSP so it will prevent the styles from being loaded
on production.

Updated `release/app` package.json and lock file to use the appropriate
name.

Updated `src/main/ipc/settings.ts` to use the ready event on the app to
check the recordings location and fixed the `mkdir` path callback where
it threw an error on no error parameter.

Updated `config/allowedOrigins.ts` from js to TypeScript to specify the
type to be inferred when used in other parts that it is just an array of
strings.

Updated `README.md` from the template to just a basic description.