Fix os target for publish workflow
Electron-builder cannot build the windows platform on linux without
using Wine, which will fail the build. Changed it to use windows.

Also added the `GITHUB_TOKEN` environment variable for that step.

Updated the `release/app/package.json` version, which is what
Electron-buider uses to create the tag for the release.